Metadata Files

SeisBase.read_meta! — Function

read_meta!(S, fmt, filestr [, keywords])

Generic wrapper for reading channel metadata (i.e., instrument parameters, responses). Reads metadata in file format fmt matching file pattern filestr into S.

This function is fully described in the official documentation at https://SeisBase.readthedocs.io/ under subheading Metadata Files.

Arguments

fmt::String: Lowercase string describing the file format.

fpat::Union{String, Vector{String}}: Read files with names matching pattern $filepat$. Supports wildcards.

For information on keyword arguments, see SeisBase.KW

See also: SeisBase.KW, get_data, read_data

SeisBase.read_meta — Function

S = read_meta(fmt, filestr [, keywords])

Creates a new SeisData object then calls read_meta! to read metadata in file format fmt matching file pattern filestr into S.

See also: read_meta!

Supported File Formats

File Format	String

Dataless SEED	dataless

FDSN Station XML	sxml

SACPZ	sacpz

SEED RESP	resp

Warning: Dataless SEED, SACPZ, and RESP files must be Unix text files; DOS text files, whose lines end in "\r\n", will not read properly. Convert with dos2unix or equivalent Windows Powershell commands.

Supported Keywords

KW	Used By	Type	Default	Meaning

memmap	all	Bool	false	use Mmap.mmap to buffer file?

msr	sxml	Bool	false	read full MultiStageResp?

s	all	TimeSpec		Start time

t	all	TimeSpec		Termination (end) time

units	resp	Bool	false	fill in MultiStageResp units?

dataless

v	all	Integer	0	verbosity

Note: mmap=true improves read speed for ASCII formats but requires caution. Julia language handling of SIGBUS/SIGSEGV and associated risks is unknown and undocumented.