SeisHDF

This submodule contains dedicated support for seismic subformats of the HDF5 file format.

Additional Functions

SeisBase.SeisHDF.asdf_waux — Function

asdf_waux(hdf_out, path, X)

Write X to AuxiliaryData/path in hdf_out. If an object already exists at AuxiliaryData/path, it will be deleted and overwritten with X.

SeisBase.SeisHDF.asdf_rqml — Function

(H,R) = asdf_rqml(fpat::String)

Read QuakeXML (qml) from ASDF file(s) matching file string pattern fpat. Returns:

H, Array{SeisHdr,1}

R, Array{SeisSrc,1}

SeisBase.SeisHDF.asdf_wqml — Function

asdf_wqml(fname, SHDR::Array{SeisHdr,1}, SSRC::Array{SeisSrc,1}[, KWs])
asdf_wqml(fname, H::SeisHdr, R::SeisSrc[, KWs])

Write QuakeXML (qml) to "QuakeML/" dataset in ASDF file fname from SHDR and SSRC.

asdf_wqml(fname, evt::SeisEvent[, KWs])
asdf_wqml(fname, evt::Array{SeisEvent,1}[, KWs])

As above, for the :hdr and :source fields of evt.

KW	Type	Default	Meaning

ovr	Bool	false	overwrite QML in existing ASDF file? [1]

v	Integer	0	verbosity

By default, data are appended to the existing contents of "QuakeML/".

Warning

To write data from R ∈ SSRC, it must be true that R.eid == H.id for some H ∈ SHDR.

See also: write_qml

SeisBase.SeisHDF.read_asdf_evt — Function

EventCat = read_asdf_evt(filestr, event_id::Union{String, Regex}[, KWs])

Read data in seismic HDF5 format with ids matching event_id from files matching pattern filestr. Returns an array of SeisEvent structures.

EventCat = read_asdf_evt(filestr, [, KWs])

Read data in seismic HDF5 format from files matching pattern filestr into SeisEvent structures. Matches any event ID in any matching file.

KW	Type	Default	Meaning

msr	Bool	true	read full (MultiStageResp) instrument resp?

v	Integer	0	verbosity

See also: TimeSpec, parsetimewin, read_data, read_hdf5

SeisBase.SeisHDF.scan_hdf5 — Function

scan_hdf5(hdf)

Scan HDF5 archive hdf and return station names with waveform data contained therein as a list of Strings formatted "nn.sssss" (network.station).

scan_hdf5(hdf, level="trace")

Scan HDF5 archive hdf and return channel names with waveform data contained therein as a list of Strings formatted "nn.sssss.ll.ccc" (network.station.location.channel).